+++ src/components/PDFBuilderComponents/PDFSkills.jsx
@@ -1,6 +1,9 @@
 import React from "react";
 import { Image } from "@react-pdf/renderer";
 import styled from "@react-pdf/styled-components";
+import angularBadge from "../../assets/images/angularBadge.png";
+import nodeBadge from "../../assets/images/nodeBadge.png";
+import cssBadge from "../../assets/images/cssBadge.png";
 
 const Root = styled.View`
   background-color: #181626;
@@ -16,7 +19,7 @@ const Header = styled.Text`
 const SubHeader = styled.Text`
   color: #fff;
   font-size: 7.5px;
-  margin-bottom: 20px;
+  margin-bottom: 8px;
 `;
 
 const Li = styled.Text`
@@ -41,10 +44,27 @@ const LiWrapper = styled.View`
   margin-bottom: 4px;
 `;
 
+const Badge = styled.Image`
+  width: 22%;
+`;
+
+const Flex = styled.View`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  margin-top: 10px;
+  margin-bottom: 10px;
+`;
+
 export function PDFSkills({ skills }) {
   return (
     <Root>
       <Header>SKILLS</Header>
+      <Flex>
+        <Badge src={angularBadge} />
+        <Badge src={nodeBadge} />
+        <Badge src={cssBadge} />
+      </Flex>
       <SubHeader>LANGUAGES - FRAMEWORKS - LIBRARIES</SubHeader>
       {skills.map((skill) => {
         return (
