+++ src/mock/mock.json
@@ -126,6 +126,24 @@
       "till": "01-06-2005",
       "level": "MBO",
       "certificate": true
+    },
+    {
+      "id": "3",
+      "institute": "Udemy",
+      "name": "React course",
+      "from": "01-09-2001",
+      "till": "01-06-2005",
+      "level": "Advanced",
+      "certificate": true
+    },
+    {
+      "id": "4",
+      "institute": "Frontend Masters",
+      "name": "Frontend webdevelopment",
+      "from": "01-09-2001",
+      "till": "01-06-2005",
+      "level": "Advanced",
+      "certificate": false
     }
   ]
 }
